refactor(ipc): move SharedMemoryManager to core/IPC/
Move SharedMemoryManager.h/.cpp, _macOS, _Windows from core/DataStruct/
to core/IPC/ to consolidate all IPC-related code. Update include paths in
Producer.cpp, main.cpp, TCMT.vcxproj, and CMakeLists.txt.
