+++ src/core/IPC/SharedMemoryManager.cpp
