+++ src/core/IPC/SharedMemoryManager_Windows.cpp
