+++ src/core/IPC/SharedMemoryManager_macOS.cpp
